Refactoring AMPArticle->transformArticleHeader
This commit extracts an AMPHeader object from the method
and moves associated code inside.

As part of the refactoring step, pulling out an AMPLogo object as well.

At the moment, this is a collection point from the three attributes of
URL, Width, and Height and can be used later to house scaling logic and
accept raw values.